% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{sssverse_deps}
\alias{sssverse_deps}
\title{List all sssverse dependencies}
\usage{
sssverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
sssverse packages.}
}
\description{
List all sssverse dependencies
}
