% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{sssverse_update}
\alias{sssverse_update}
\title{Update sssverse packages}
\usage{
sssverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
sssverse packages.}
}
\description{
This will check to see if all sssverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
sssverse_update()
}
}
